requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
aiohttp>=3.9.0
python-dotenv>=1.0.1
jinja2>=3.1.0

